Added ./Analytics
Added an analytics submission page that takes a setting and stub text
input, then returns information about the corresponding page’s stats.
On the backend, this makes a request to GoogleAnalytics API v4 for this
information, then converts that information into a more usable object
before returning it to the user.