/Analytics GoogleCharts Update
Managed to incorporate google charts without the third party
React-Google-Charts module. There are still two issues unresolved.

1. Google loader is called more than once on load. This is most likely
due to the redundancy of having the loader in the server.js as well as
the third party module having the loader.

2. Google Charts takes time to load, so if a request is made to quickly
when the page loads, the chart will not load. The error was prevented
with a check for google.visualizations, however this does not
automatically render the chart when google charts is ready.

Other change: small change to naming in ReactGoogleCharts file, for
compatibility with this new Google Charts graph, as well as clarity in
whats being exported.