fix: Add a few additional redirects neeed by BAAS
The Bulletin of the American Astronomical Society (BAAS) just replaced their
old WordPress site with PubPub, and there are a few URLs out in the wild that
we want to preserve by establishing redirects. A couple of the new entries
that we need are because PubPub doesn't accept URLs with trailing slashes for
its established names. One of them is for a URL slug that apparently got
changed at some point.