feat(sdk): move everything around until imports work
This is not a good solution. I'm mainly trying to avoid importing too much from the client, but it will be too hard to completely avoid that in the future I think, may need to come up with a better solution or this may prove to be fruitless because of it.